requirements.txt
pandas
pillow
requests
scikit-learn
streamlit